Implemented Audit Logging and updated UI
Added the Audit Logging Plugin and configured it to track changes to the Patient entity. Made substantial changes to the User Interface (UI), including color updates and form changes. Additionally, multiple domain classes have been updated and improved along with tweaks to Gradle settings and added tests for the new AuditTrail domain. This commit also contains minor modification to the project's Grails configuration, which is expected to increase efficiency.