Assign main clinic and calculate matchId for new patient
In the save method of PatientController, two enhancements are introduced. Firstly, the main clinic is assigned to a patient while saving. Secondly, the matchId is calculated for the new patient record. If no such records exist yet, the matchId will be 1; otherwise, it will be incremented from the latest matchId.